Changes in response to review by L.J.
The suggested changes have been implemented. The major change is to the function name, with minor updates to the documentation.

Note 1. The check is for non-alpha characters (not non-alphanumeric) these should not be in formal taxon names. They are expected to help flag cases where informal names to distinguish informal morphotypes eg. Moth_sp.1, Moth_sp.2.

Note 2. An example of a common prefix has been added to illustrate how this similarity can produce spurious matches. When I was developing the function, things like Protogondollela and Progondollela would be matched so I gave a way to exclude by prefix and suffix.

Note 3. I have retained a similarity threshold of 80% for now. >90 similarity will substantially reduce the number of irrelevant flags, but I have generally found that >80 will ensure that the returned results will contain all potential synonyms. A bit of a trade off here, but I think a less stringent threshold is a bit safer.