Updates in response to review by K.E and L.J.
Function has been updated so that the synonyms are presented first, then any non-letter checks if requested by the user. Frequencies of synonyms in a pair are now reported.

Removal of the pref/suff matching was suggested as a possible means of simplifying the function. I have held off on doing this for now as I think this functionality does have merit for checking taxonomic data, so I have attempted to clarify their usage more comprehensively in the documentation. I have removed the end letter matching, however, as I concluded that it was unnecessary.

Finally, I have kept the data.frame output largely the same for now as opposed to a structure that would allow grouping of synonyms, where there are three or more names all similar to each other. I have encountered this situation in earlier versions of this function, but I can't see a convenient way of doing this such that that output is easy for the user to parse. The best I have is alphabetical organisation of the names so that triplets/quads ect of similar names should appear in adjacent rows as sets (already implemented).

I've done my best to check lint and code coverage before submission, but apologies if I have missed anything. I am completely new to this sort of collaborative coding