Addition of function: tax_spellcheck (#30)
* Addition of tax_spellcheck

tax_spellcheck provides a quick means of identifying potential spelling variations of the same taxon name. This flagging enables user correction of their data. Function addition was invited by developer Lewis Jones.

* Update tax_spellcheck.R

* Update tax_spellcheck.R

* Minor bug fixes, test coverage boosted to 90%

* Delete codecov.yml

* Re-add codecov

* Changes in response to review by L.J.

The suggested changes have been implemented. The major change is to the function name, with minor updates to the documentation.

Note 1. The check is for non-alpha characters (not non-alphanumeric) these should not be in formal taxon names. They are expected to help flag cases where informal names to distinguish informal morphotypes eg. Moth_sp.1, Moth_sp.2.

Note 2. An example of a common prefix has been added to illustrate how this similarity can produce spurious matches. When I was developing the function, things like Protogondollela and Progondollela would be matched so I gave a way to exclude by prefix and suffix.

Note 3. I have retained a similarity threshold of 80% for now. >90 similarity will substantially reduce the number of irrelevant flags, but I have generally found that >80 will ensure that the returned results will contain all potential synonyms. A bit of a trade off here, but I think a less stringent threshold is a bit safer.

* Updates in response to code review by L.J

Output changed to report any non-letter characters, as well as any flagged synonyms.

Streamlined argument checking at the start of the function.

Documentation and code fully updated so that similarity is in range 0-100, rather than 0-1.

Updates for better handling of NA and blank strings.

Fix to potential bug during removal of equivalent pairs (i.e xy, yx rows of names).

Tetrapods dataset now used in the example

* Update R/tax_check.R

Co-authored-by: Lewis A. Jones <41071747+LewisAJones@users.noreply.github.com>

* Update R/tax_check.R

Co-authored-by: Lewis A. Jones <41071747+LewisAJones@users.noreply.github.com>

* Update R/tax_check.R

Co-authored-by: Lewis A. Jones <41071747+LewisAJones@users.noreply.github.com>

* Update R/tax_check.R

Co-authored-by: Lewis A. Jones <41071747+LewisAJones@users.noreply.github.com>

* Update R/tax_check.R

Co-authored-by: Lewis A. Jones <41071747+LewisAJones@users.noreply.github.com>

* Update R/tax_check.R

Co-authored-by: Lewis A. Jones <41071747+LewisAJones@users.noreply.github.com>

* Update R/tax_check.R

Co-authored-by: Lewis A. Jones <41071747+LewisAJones@users.noreply.github.com>

* Update R/tax_check.R

Co-authored-by: Lewis A. Jones <41071747+LewisAJones@users.noreply.github.com>

* Update R/tax_check.R

Co-authored-by: Lewis A. Jones <41071747+LewisAJones@users.noreply.github.com>

* Update R/tax_check.R

Co-authored-by: Lewis A. Jones <41071747+LewisAJones@users.noreply.github.com>

* Update R/tax_check.R

Co-authored-by: Lewis A. Jones <41071747+LewisAJones@users.noreply.github.com>

* Update R/tax_check.R

Co-authored-by: Lewis A. Jones <41071747+LewisAJones@users.noreply.github.com>

* Update R/tax_check.R

Co-authored-by: Kilian Eichenseer <40391500+KEichenseer@users.noreply.github.com>

* Update R/tax_check.R

Co-authored-by: Kilian Eichenseer <40391500+KEichenseer@users.noreply.github.com>

* Update R/tax_check.R

Co-authored-by: Kilian Eichenseer <40391500+KEichenseer@users.noreply.github.com>

* Update R/tax_check.R

Co-authored-by: Kilian Eichenseer <40391500+KEichenseer@users.noreply.github.com>

* Update R/tax_check.R

Co-authored-by: Kilian Eichenseer <40391500+KEichenseer@users.noreply.github.com>

* Updates in response to review by K.E and L.J.

Function has been updated so that the synonyms are presented first, then any non-letter checks if requested by the user. Frequencies of synonyms in a pair are now reported.

Removal of the pref/suff matching was suggested as a possible means of simplifying the function. I have held off on doing this for now as I think this functionality does have merit for checking taxonomic data, so I have attempted to clarify their usage more comprehensively in the documentation. I have removed the end letter matching, however, as I concluded that it was unnecessary.

Finally, I have kept the data.frame output largely the same for now as opposed to a structure that would allow grouping of synonyms, where there are three or more names all similar to each other. I have encountered this situation in earlier versions of this function, but I can't see a convenient way of doing this such that that output is easy for the user to parse. The best I have is alphabetical organisation of the names so that triplets/quads ect of similar names should appear in adjacent rows as sets (already implemented).

I've done my best to check lint and code coverage before submission, but apologies if I have missed anything. I am completely new to this sort of collaborative coding

* Examples update

Simpler example without prefixes and suffixes added

* Update tax_check.R

* Update tax_check.R

* pref-suff clarification

Documentation for pref and suff has been revised. I agree totally with Killian's suggested argument descriptions and they made me realise that I had been thinking about the prefixes and suffixes in slightly the wrong way. Hopefully the revisions will now make their precise functionality clear.

I have re-added the brachiopods example that Lewis previously requested removal of, as this properly shows the effect of using prefixes and suffixes during synonym checking compared to the tetrapod dataset. It is a relatively small call to the PBDB API and runs pretty quickly, but I have wrapped the example in {donotrun} to prevent CRAN from throwing a hissy fit if the example runtime exceeds their regulations

* Update tax_check.R

* tax_check revamp

* fix lint issues and error in bin_lat documentation

* Corrections to tax_check

Co-authored-by: Lewis A. Jones <41071747+LewisAJones@users.noreply.github.com>
Co-authored-by: Kilian Eichenseer <40391500+KEichenseer@users.noreply.github.com>
Co-authored-by: Kilian Eichenseer <kilian.eichenseer@gmail.com>